Fixed issues: Manage time functionality for per form.

+++ app/src/main/java/com/rf/geolgy/ui/main/activity/DashBoardActivity.kt
@@ -121,7 +121,7 @@ class DashBoardActivity : BaseActivity<ActivityDashboardBinding>(R.layout.activi
                     validFrom = validFrom,
                     validTo = validTo,
                     gstNumber = gstNumber,
-                    expireInHours = viewDataBinding?.spinnerExpire?.text.toString()
+                    expireInHours = viewDataBinding?.spinnerExpire?.text.toString().toInt()
                 )
                 createChallanAPI(createChallanRequest, gstNumber)
             } else {
